requirements.txt
fastapi==0.115.12
uvicorn[standard]>=0.20.0
gunicorn>=21.0.0
azure-ai-inference>=1.0.0b9
azure-core>=1.30.0
python-multipart==0.0.20
pydantic==1.10.12
aiofiles>=23.0.0
PyPDF2>=3.0.0
python-docx>=1.0.0
python-dotenv>=1.0.0
